+++ src/main/java/com/floney/floney/book/service/BookLineServiceImpl.java
@@ -149,7 +149,7 @@ public class BookLineServiceImpl implements BookLineService {
         return bookLineRepository.dayIncomeAndOutcome(bookKey, dates);
     }
 
-    private Map<String, Float> totalExpense(String bookKey, DatesDuration dates) {
+    private Map<String, Double> totalExpense(String bookKey, DatesDuration dates) {
         return bookLineRepository.totalExpenseByMonth(bookKey, dates);
     }
 }
