+++ src/main/java/com/floney/floney/book/service/BookServiceImpl.java
@@ -41,7 +41,7 @@ import static com.floney.floney.common.constant.Subscribe.SUBSCRIBE_MAX_BOOK;
 public class BookServiceImpl implements BookService {
 
     private static final int ONLY_OWNER_COUNT = 1;
-    private static final float DEFAULT_BUDGET = 0f;
+    private static final double DEFAULT_BUDGET = 0.0;
 
     private final BookRepository bookRepository;
     private final BookUserRepository bookUserRepository;
@@ -306,9 +306,9 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public Map<Month, Float> getBudgetByYear(String bookKey, String firstDate) {
+    public Map<Month, Double> getBudgetByYear(String bookKey, String firstDate) {
         LocalDate date = LocalDate.parse(firstDate);
-        Map<Month, Float> monthlyMap = getInitBudgetFrame();
+        Map<Month, Double> monthlyMap = getInitBudgetFrame();
 
         List<BudgetYearResponse> savedBudget = bookRepository.findBudgetByYear(bookKey, DateFactory.getYearDuration(date));
 
@@ -386,8 +386,8 @@ public class BookServiceImpl implements BookService {
         return !bookUser.isInactive() && bookUserRepository.countByBookExclusively(bookUser.getBook()) == ONLY_OWNER_COUNT;
     }
 
-    private Map<Month, Float> getInitBudgetFrame() {
-        Map<Month, Float> monthlyMap = new LinkedHashMap<>();
+    private Map<Month, Double> getInitBudgetFrame() {
+        Map<Month, Double> monthlyMap = new LinkedHashMap<>();
         for (Month month : Month.values()) {
             monthlyMap.put(month, DEFAULT_BUDGET);
         }
